feat: notify transfer to reward auth wallet instead
