+++ scripts/create_pool/config.json
@@ -1,6 +1,6 @@
 {
-  "RPC_ENDPOINT_URL": "https://api-mainnet-beta.renec.foundation:8899",
-  "REDEX_PROGRAM_ID": "7rh7ZtPzHqdY82RWjHf1Q8NaQiWnyNqkC48vSixcBvad",
+  "RPC_ENDPOINT_URL": "https://api-testnet.renec.foundation:8899",
+  "REDEX_PROGRAM_ID": "EwGzCLvNKeLGAP9QymNgneJBd1gmJH75WnhZKo5Chy4W",
   "PROTOCOL_FEE_RATE": 300,
   "FEE_TIERS": [
     {
