+++ scripts/create_pool/deployed.json
@@ -1,1 +1,1 @@
-{"REDEX_CONFIG_PUB": ""}
+{ "REDEX_CONFIG_PUB": "" }
