+++ scripts/helper/02_transfer_token_to_ata.ts
@@ -1,44 +1,14 @@
 import { PublicKey, Keypair, Transaction } from "@solana/web3.js";
 import {
-  buildWhirlpoolClient,
-  WhirlpoolIx,
-  toTx,
-  WhirlpoolContext,
-  InitializeRewardParams,
-} from "@renec/redex-sdk";
-import { Whirlpool } from "@renec/redex-sdk/src";
-import {
   getTokenMintInfo,
   loadProvider,
   loadWallets,
 } from "../create_pool/utils";
 import { deriveATA } from "@orca-so/common-sdk";
-import { AnchorProvider } from "@project-serum/anchor";
-import { Token, TOKEN_PROGRAM_ID, u64 } from "@solana/spl-token";
 import Decimal from "decimal.js";
-import { DecimalUtil, Percentage } from "@orca-so/common-sdk";
+import { DecimalUtil } from "@orca-so/common-sdk";
+import { transfer } from "../create_pool/utils/token";
 
-export async function transfer(
-  provider: AnchorProvider,
-  source: PublicKey,
-  destination: PublicKey,
-  amount: number | u64
-) {
-  const tx = new Transaction();
-  tx.add(
-    Token.createTransferInstruction(
-      TOKEN_PROGRAM_ID,
-      source,
-      destination,
-      provider.wallet.publicKey,
-      [],
-      amount
-    )
-  );
-  return provider.sendAndConfirm(tx, [], { commitment: "confirmed" });
-}
-
-//usage: yarn set_pool_reward <pool_address> <reward_token_mint> <reward_index?>
 async function main() {
   // Load context
   const wallets = loadWallets();
